A set of maintenance patches to enable darktable to process images of (more or less) arbitrary
size. We can take advantage of the already long-established tiling mechanism. The only thing that
needed fixing is the type of variable used as index into image buffers. Our code used a lot of
type int (int32_t), which is now replaced by size_t. (No guarantee that I catched all cases - please
report problems)

To process huge images (e.g. stitched panoramas coming from Hugin) you need to have a look at the
settings of config parameter "host memory limit" (core options). It needs to be high enough to let
darktable take full advantage of all your system memory but not so high that your system runs into
HDD-swapping. You will also need to increase config parameter "maximum_number_tiles" (darktablerc)
to something like 10000 (the new default value).

If you are running darktable on a 32-bit system, don't even try to import anything bigger than an ordinary
RAW.

+++ src/common/imageio_png.c
@@ -189,7 +189,7 @@ dt_imageio_open_png(
     return DT_IMAGEIO_CACHE_FULL;
   }
 
-  buf = dt_alloc_align(16, width*height*3*(bpp < 16 ? 1 : 2));
+  buf = dt_alloc_align(16, (size_t)width*height*3*(bpp < 16 ? 1 : 2));
   if(!buf)
   {
     fclose(image.f);
@@ -205,11 +205,11 @@ dt_imageio_open_png(
     return DT_IMAGEIO_FILE_CORRUPTED;
   }
 
-  for (uint32_t j = 0; j < height; j++)
+  for (size_t j = 0; j < height; j++)
   {
-    if(bpp < 16) for(uint32_t i=0; i<width; i++)
+    if(bpp < 16) for(size_t i=0; i<width; i++)
         for(int k=0; k<3; k++) mipbuf[4*(j * width + i) + k] = buf[3*(j * width + i) + k]*(1.0f/255.0f);
-    else for(uint32_t i=0; i<width; i++)
+    else for(size_t i=0; i<width; i++)
         for(int k=0; k<3; k++) mipbuf[4*(j * width + i) + k] = (256.0f*buf[2*(3*(j * width + i) + k)] + buf[2*(3*(j * width + i) + k) + 1])*(1.0f/65535.0f);
   }
