+++ src/develop/imageop.h
@@ -155,8 +155,8 @@ typedef struct dt_iop_module_so_t
   int  (*process_cl)      (struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece, void *i, void *o, const struct dt_iop_roi_t *roi_in, const struct dt_iop_roi_t *roi_out);
   int  (*process_tiling_cl)      (struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece, void *i, void *o, const struct dt_iop_roi_t *roi_in, const struct dt_iop_roi_t *roi_out, const int bpp);
 
-  int (*distort_transform) (struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece, float *points, int points_count);
-  int (*distort_backtransform) (struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece, float *points, int points_count);
+  int (*distort_transform) (struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece, float *points, size_t points_count);
+  int (*distort_backtransform) (struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece, float *points, size_t points_count);
 }
 dt_iop_module_so_t;
 
@@ -326,9 +326,9 @@ typedef struct dt_iop_module_t
    * points is an array of float {x1,y1,x2,y2,...}
    * size is 2*points_count */
   /** points before the iop is applied => point after processed */
-  int (*distort_transform) (struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece, float *points, int points_count);
+  int (*distort_transform) (struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece, float *points, size_t points_count);
   /** reverse points after the iop is applied => point before process */
-  int (*distort_backtransform) (struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece, float *points, int points_count);
+  int (*distort_backtransform) (struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece, float *points, size_t points_count);
 
   /** Key accelerator registration callbacks */
   void (*connect_key_accels)(struct dt_iop_module_t *self);
@@ -503,8 +503,8 @@ static inline void dt_iop_alpha_copy(const void *ivoid, void *ovoid, const int w
 #endif
   for(int j=0; j<height; j++)
   {
-    const float *in  = ((const float *)ivoid)+4*width*j+3;
-    float *out = ((float *)ovoid)+4*width*j+3;
+    const float *in  = ((const float *)ivoid)+(size_t)4*width*j+3;
+    float *out = ((float *)ovoid)+(size_t)4*width*j+3;
     for(int i=0; i<width; i++)
     {
       *out = *in;
